third_party/WebKit/LayoutTests/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_memory_cached-expected.txt

Description

This test validates that a memory cached resource appears in the buffer once.

  

PASS There should be only one entry 
PASS http://127.0.0.1:8000/w3c/webperf/resources/generate_resource.php?type=image&cacheable=1&id=cached is expected to have initiatorType img 
PASS requestStart should non-zero on the same-origin request 
PASS responseEnd should not be before startTime 
PASS duration should not be negative 
PASS There should be only one entry 
PASS http://localhost:8000/w3c/webperf/resources/generate_resource.php?type=image&cacheable=1&id=cached should be present 
PASS requestStart should be 0 on the cross-origin request 
